reorganized components
