+++ client/components/GameTable/PlayerHand/Card.js
@@ -1,15 +1,15 @@
 /*eslint-disable complexity */
 import React from 'react'
 import {connect} from 'react-redux'
-import {selectCard} from '../store/selectedCard'
+import {selectCard} from '../../../store/selectedCard'
 import {Button} from 'semantic-ui-react'
-import TradeModal from './TradeModal'
-import checkTrading from '../../gameLogic/checkTrading'
-import {selectAction} from '../store/selectedAction'
-import {playCard} from '../store/playCard'
-import checkWonderLevelOneTrading from '../../gameLogic/checkWonderLevelOneTrading'
-import checkWonderLevelTwoTrading from '../../gameLogic/checkWonderLevelTwoTrading'
-import checkWonderLevelThreeTrading from '../../gameLogic/checkWonderLevelThreeTrading'
+import TradeModal from '../Trade/TradeModal'
+import checkTrading from '../../../../gameLogic/checkTrading'
+import {selectAction} from '../../../store/selectedAction'
+import {playCard} from '../../../store/playCard'
+import checkWonderLevelOneTrading from '../../../../gameLogic/checkWonderLevelOneTrading'
+import checkWonderLevelTwoTrading from '../../../../gameLogic/checkWonderLevelTwoTrading'
+import checkWonderLevelThreeTrading from '../../../../gameLogic/checkWonderLevelThreeTrading'
 
 class Card extends React.Component {
   render() {
