+++ client/components/GameTable/Trade/TradeItem.js
