+++ client/components/Lobby/GameRoomSnapshot.js
