+++ client/components/NavBar/QuitMessage.js
