+++ client/components/WelcomePage/Welcome.js
